Factor out exp/grade calculation code to a sperate file
